+++ pkg/core/config/downloader.go
@@ -1,0 +1,7 @@
+package config
+
+import "github.com/simpleg-eu/cuplan-core/pkg/core"
+
+type Downloader interface {
+	Download(host string, stage string, environment string, component string) core.Result[[]byte, core.Error]
+}
