+++ pkg/core/empty.go
@@ -1,0 +1,3 @@
+package core
+
+type Empty struct{}
