+++ pkg/core/error.go
@@ -1,4 +1,4 @@
-package pkg
+package core
 
 type Error struct {
 	errorKind string
