+++ pkg/core/error_kind.go
@@ -1,4 +1,5 @@
-package pkg
+package core
 
 const CommandFailure string = "command_failure"
 const SerializationFailure string = "serialization_failure"
+const MissingFilePath string = "missing_file_path"
