+++ pkg/core/result.go
@@ -1,4 +1,4 @@
-package pkg
+package core
 
 type Result[OkType any, ErrorType any] struct {
 	ok   OkType
