+++ pkg/core/secret/bitwarden_provider_test.go
@@ -1,7 +1,7 @@
 package secret
 
 import (
-	core "github.com/simpleg-eu/cuplan-core/pkg"
+	"github.com/simpleg-eu/cuplan-core/pkg/core"
 	"github.com/stretchr/testify/assert"
 	"os"
 	"testing"
