+++ pkg/core/secret/provider.go
@@ -1,6 +1,6 @@
 package secret
 
-import core "github.com/simpleg-eu/cuplan-core/pkg"
+import "github.com/simpleg-eu/cuplan-core/pkg/core"
 
 type Provider interface {
 	Get(string) core.Result[string, core.Error]
